new flat surface algo still missing boundries

+++ GenClasses/Tiles.py
@@ -91,7 +91,7 @@ class tileGatherer():
                                         lat,lon =  self.converter.PixelXYToLatLongOSM(x_pixal_world,y_pixal_world,15)
                                         color = filled_tiles_matrix[x][y][i][j]
                                         maerc_lat,maerc_lon = transformer.transform(lat, lon)
-                                        height = -10000 + ((color[0] * 256 * 256 + color[1] * 256 + color[2]) * 0.1)
+                                        height = float(-10000 + ((color[0] * 256 * 256 + color[1] * 256 + color[2]) * 0.1))
                                         temp2.append({'lat':lat,'lon':lon,'Mercator':[maerc_lat,maerc_lon] , 'color':color ,'world_pixal_X':x_pixal_world,'world_pixal_Y':y_pixal_world , 'pixal_X':x_pixal,'pixal_Y':y_pixal,'height':height})
                                         if Point(maerc_lat,maerc_lon).within(Polygon(self.userMarker.square)):
                                                 data = Points([lat,lon], [maerc_lat,maerc_lon] , [x_pixal,y_pixal] , [x_pixal_world,y_pixal_world] , height , color)
